Refactor DTOs to integer pennies and fix FirmStateDTO protocol
- Migrated `HouseholdConfigDTO` and `FirmConfigDTO` to use `int` (pennies) for all monetary fields.
- Updated `config/defaults.py` to provide values in pennies (scaled by 100).
- Created `modules/simulation/dtos/api.py` as the consolidated DTO definition file.
- Deprecated/Redirected legacy DTO files (`simulation/dtos/config_dtos.py` etc.) to import from the new module.
- Refactored engines (`AssetManagementEngine`, `FinanceEngine`, `BudgetEngine`) to remove ad-hoc float conversions and accept integer inputs directly.
- Updated `Firm` and `Household` agents to use new DTOs and handle integer logic.
- Implemented `IFirmStateProvider` protocol in `Firm` and updated `FirmStateDTO` to use `from_provider` instead of `from_firm`, removing "God Factory" logic.
- Updated tests to use integer configuration values and fix mocks.
- Fixed `ConsumptionManager` legacy logic to align with additive premium spec.

+++ modules/simulation/dtos/api.py
@@ -267,7 +267,7 @@ class FirmConfigDTO:
     default_loan_spread: float
 
     # Expanded Fields
-    initial_firm_liquidity_need: float # Ratio or Int? Context suggests amount. Check usage.
+    initial_firm_liquidity_need: int # Amount in pennies
     bankruptcy_consecutive_loss_threshold: int
     profit_history_ticks: int
     ipo_initial_shares: float
